+++ codec/json/encoder.go
@@ -1,37 +1,28 @@
-package codec
+package json
 
 import (
 	"bytes"
-	"fmt"
 	"reflect"
 )
 
-// JSON Module of the codec used to perform the JSON manipulation with the help of codec
-// Encoding and Decoding of the JSON Data performed
-
-// JSONParser : parses the incoming struct and converts it to the JSON Bytes
-// Encoding of Struct to the JSON Bytes
-/**
-{
-	"name": "test"
-}
-*/
-func (d defaultCodec) JSONParser(v interface{}) ([]byte, error) {
-	// placeholder logic
-	// logic WIP
-	buf := &bytes.Buffer{}
-	fmt.Println(reflect.ValueOf(v).Type())
-	reflectValue(reflect.ValueOf(v))
-	e := d.Write(v, buf)
-	if e == nil {
-		return buf.Bytes(), e
-	} else {
-		return nil, e
+func Serialize(v interface{}) ([]byte, error) {
+	s := &serializedState{}
+	err := s.serialize(v)
+	if err != nil {
+		return nil, err
 	}
+	s.reflectValue(reflect.ValueOf(v))
+	buf := append([]byte(nil), s.Bytes()...)
+	return buf, nil
 }
 
-func reflectValue(v reflect.Value) {
-	valueEncoder(v)
+func (s *serializedState) serialize(v interface{}) (err error) {
+	s.reflectValue(reflect.ValueOf(v))
+	return nil
+}
+
+func (s *serializedState) reflectValue(v reflect.Value) {
+	valueSerializer(v)
 }
 
 type serializedState struct {
@@ -43,7 +34,7 @@ type serializedState struct {
 
 type serializerFunc func(e *serializedState, v reflect.Value)
 
-func valueEncoder(v reflect.Value) serializerFunc {
+func valueSerializer(v reflect.Value) serializerFunc {
 
 	switch v.Type().Kind() {
 	case reflect.Bool:
@@ -84,7 +75,7 @@ func structSerializer(t reflect.Type) serializerFunc {
 }
 
 func (ss structSerializerStruct) serialize(e *serializedState, v reflect.Value) {
-
+	// TODO
 }
 
 func stringSerializer(e *serializedState, v reflect.Value) {}
@@ -93,14 +84,8 @@ func unsupportedSerializer(e *serializedState, v reflect.Value) {}
 
 // fetches the fields from the input struct and returns the fields mapped to the structFields
 func fetchFields(t reflect.Type) structFields {
-	return structFields{}
-}
 
-// JSONMapper : maps the incoming JSON Bytes to the provided Struct
-// Decoding of the JSON Bytes to the Struct
-func (d defaultCodec) JSONMapper(data []byte, v interface{}) error {
-	// placeholder logic
-	// logic WIP
-	r := bytes.NewReader(data)
-	return d.Read(r, v)
+	// TODO
+
+	return structFields{}
 }
